Welcome to drafft.ink. It's fast, it's open, and it took a week of my life. Enjoy.
This is the result of a personal quest to create a high-performance,
genuinely self-hostable whiteboard after failing to get support from
closed/semi-open alternatives.

The project is built entirely in Rust, uses egui for the cross-platform
GUI, and WebGPU for accelerated rendering. The primary license is AGPLv3
to ensure that any service built upon this code must remain open for the
end-user.

This commit contains:

- The entire Rust core application.
- The WebSocket server implementation.
- All frontend integration (native and WASM).

Go forth and improve!